day 17 solve!
I can proudly say that I made it further than last year without needing to look up the solution. This one was a doozy for me-
but in retrospect the intuition for part 2 is fairly obvious ( I just missed that there are multiple valid choices when selecting
a digit, which alters the state down the line). Genuinely excited to have figured this one out on my own.