2025 Day 2
late to the party this time around but oh well.. finals are tough

this is a very brute-forcey solution. It should be possible to look at the patterns of starting digits
and identify where repeated sequences of digits are within each range rather than checking each number in the range.